refactor: Move extension traits for snapshot into snapshot plugin
