+++ src/lib.rs
@@ -4,11 +4,7 @@
 #![allow(clippy::type_complexity)] // Suppress warnings around Query
 
 use bevy::{
-    ecs::{
-        component::Mutable,
-        entity::MapEntities,
-        schedule::{ExecutorKind, LogLevel, ScheduleBuildSettings, ScheduleLabel},
-    },
+    ecs::schedule::{ExecutorKind, LogLevel, ScheduleBuildSettings, ScheduleLabel},
     input::InputSystem,
     platform::collections::HashMap,
     prelude::*,
@@ -30,7 +26,7 @@ pub(crate) mod time;
 pub mod prelude {
     pub use crate::{
         snapshot::prelude::*, AddRollbackCommandExtension, GgrsApp, GgrsConfig, GgrsPlugin,
-        GgrsSchedule, GgrsTime, PlayerInputs, ReadInputs, Rollback, Session,
+        GgrsSchedule, GgrsTime, PlayerInputs, ReadInputs, Rollback, RollbackApp, Session,
     };
     pub use ggrs::{GgrsEvent, PlayerType, SessionBuilder};
 }
@@ -200,84 +196,10 @@ impl<C: Config> Plugin for GgrsPlugin<C> {
     }
 }
 
-/// Extension trait to add the GGRS plugin idiomatically to Bevy Apps
-pub trait GgrsApp {
-    /// Registers a component type for saving and loading from the world. This
-    /// uses [`Copy`] based snapshots for rollback.
-    fn rollback_component_with_copy<Type>(&mut self) -> &mut Self
-    where
-        Type: Component + Copy;
-
-    /// Registers a resource type for saving and loading from the world. This
-    /// uses [`Copy`] based snapshots for rollback.
-    fn rollback_resource_with_copy<Type>(&mut self) -> &mut Self
-    where
-        Type: Resource + Copy;
-
-    /// Registers a component type for saving and loading from the world. This
-    /// uses [`Clone`] based snapshots for rollback.
-    fn rollback_component_with_clone<Type>(&mut self) -> &mut Self
-    where
-        Type: Component + Clone;
-
-    /// Registers a resource type for saving and loading from the world. This
-    /// uses [`Clone`] based snapshots for rollback.
-    fn rollback_resource_with_clone<Type>(&mut self) -> &mut Self
-    where
-        Type: Resource + Clone;
-
-    /// Registers a component type for saving and loading from the world. This
-    /// uses [`reflection`](`Reflect`) based snapshots for rollback.
-    ///
-    /// NOTE: Unlike previous versions of `bevy_ggrs`, this will no longer automatically
-    /// apply entity mapping through the [`MapEntities`](`bevy::ecs::entity::MapEntities`) trait.
-    /// If you require this behavior, see [`ComponentMapEntitiesPlugin`].
-    fn rollback_component_with_reflect<Type>(&mut self) -> &mut Self
-    where
-        Type: Component + Reflect + FromWorld;
-
-    /// Registers a resource type for saving and loading from the world. This
-    /// uses [`reflection`](`Reflect`) based snapshots for rollback.
-    ///
-    /// NOTE: Unlike previous versions of `bevy_ggrs`, this will no longer automatically
-    /// apply entity mapping through the [`MapEntities`](`bevy::ecs::entity::MapEntities`) trait.
-    /// If you require this behavior, see [`ComponentMapEntitiesPlugin`].
-    fn rollback_resource_with_reflect<Type>(&mut self) -> &mut Self
-    where
-        Type: Resource + Reflect + FromWorld;
-
+/// exstension trait for [`App`] to add GGRS specific functionality
+pub trait GgrsApp: RollbackApp {
     /// Set the frequency that game updates should be performed at.
     fn set_rollback_schedule_fps(&mut self, fps: usize) -> &mut Self;
-
-    /// Adds a component type to the checksum generation pipeline using [`Hash`].
-    fn checksum_component_with_hash<Type>(&mut self) -> &mut Self
-    where
-        Type: Component + Hash;
-
-    /// Updates a component after rollback using [`MapEntities`].
-    fn update_component_with_map_entities<Type>(&mut self) -> &mut Self
-    where
-        Type: Component<Mutability = Mutable> + MapEntities;
-
-    /// Adds a resource type to the checksum generation pipeline using [`Hash`].
-    fn checksum_resource_with_hash<Type>(&mut self) -> &mut Self
-    where
-        Type: Resource + Hash;
-
-    /// Updates a resource after rollback using [`MapEntities`].
-    fn update_resource_with_map_entities<Type>(&mut self) -> &mut Self
-    where
-        Type: Resource + MapEntities;
-
-    /// Adds a component type to the checksum generation pipeline.
-    fn checksum_component<Type>(&mut self, hasher: for<'a> fn(&'a Type) -> u64) -> &mut Self
-    where
-        Type: Component;
-
-    /// Adds a resource type to the checksum generation pipeline.
-    fn checksum_resource<Type>(&mut self, hasher: for<'a> fn(&'a Type) -> u64) -> &mut Self
-    where
-        Type: Resource;
 }
 
 impl GgrsApp for App {
@@ -285,89 +207,5 @@ impl GgrsApp for App {
         self.world_mut().insert_resource(RollbackFrameRate(fps));
 
         self
-    }
-
-    fn rollback_component_with_reflect<Type>(&mut self) -> &mut Self
-    where
-        Type: Component + Reflect + FromWorld,
-    {
-        self.add_plugins(ComponentSnapshotPlugin::<ReflectStrategy<Type>>::default())
-    }
-
-    fn rollback_resource_with_reflect<Type>(&mut self) -> &mut Self
-    where
-        Type: Resource + Reflect + FromWorld,
-    {
-        self.add_plugins(ResourceSnapshotPlugin::<ReflectStrategy<Type>>::default())
-    }
-
-    fn rollback_component_with_copy<Type>(&mut self) -> &mut Self
-    where
-        Type: Component + Copy,
-    {
-        self.add_plugins(ComponentSnapshotPlugin::<CopyStrategy<Type>>::default())
-    }
-
-    fn rollback_resource_with_copy<Type>(&mut self) -> &mut Self
-    where
-        Type: Resource + Copy,
-    {
-        self.add_plugins(ResourceSnapshotPlugin::<CopyStrategy<Type>>::default())
-    }
-
-    fn rollback_component_with_clone<Type>(&mut self) -> &mut Self
-    where
-        Type: Component + Clone,
-    {
-        self.add_plugins(ComponentSnapshotPlugin::<CloneStrategy<Type>>::default())
-    }
-
-    fn rollback_resource_with_clone<Type>(&mut self) -> &mut Self
-    where
-        Type: Resource + Clone,
-    {
-        self.add_plugins(ResourceSnapshotPlugin::<CloneStrategy<Type>>::default())
-    }
-
-    fn checksum_component_with_hash<Type>(&mut self) -> &mut Self
-    where
-        Type: Component + Hash,
-    {
-        self.add_plugins(ComponentChecksumPlugin::<Type>::default())
-    }
-
-    fn update_component_with_map_entities<Type>(&mut self) -> &mut Self
-    where
-        Type: Component<Mutability = Mutable> + MapEntities,
-    {
-        self.add_plugins(ComponentMapEntitiesPlugin::<Type>::default())
-    }
-
-    fn checksum_resource_with_hash<Type>(&mut self) -> &mut Self
-    where
-        Type: Resource + Hash,
-    {
-        self.add_plugins(ResourceChecksumPlugin::<Type>::default())
-    }
-
-    fn update_resource_with_map_entities<Type>(&mut self) -> &mut Self
-    where
-        Type: Resource + MapEntities,
-    {
-        self.add_plugins(ResourceMapEntitiesPlugin::<Type>::default())
-    }
-
-    fn checksum_component<Type>(&mut self, hasher: for<'a> fn(&'a Type) -> u64) -> &mut Self
-    where
-        Type: Component,
-    {
-        self.add_plugins(ComponentChecksumPlugin::<Type>(hasher))
-    }
-
-    fn checksum_resource<Type>(&mut self, hasher: for<'a> fn(&'a Type) -> u64) -> &mut Self
-    where
-        Type: Resource,
-    {
-        self.add_plugins(ResourceChecksumPlugin::<Type>(hasher))
     }
 }
