+++ src/snapshot/mod.rs
@@ -14,6 +14,7 @@ mod resource_checksum;
 mod resource_map;
 mod resource_snapshot;
 mod rollback;
+mod rollback_app;
 mod rollback_entity_map;
 mod set;
 mod strategy;
@@ -29,6 +30,7 @@ pub use resource_checksum::*;
 pub use resource_map::*;
 pub use resource_snapshot::*;
 pub use rollback::*;
+pub use rollback_app::*;
 pub use rollback_entity_map::*;
 pub use set::*;
 pub use strategy::*;
